Refactor CLI module structure
