+++ src/padz/cli/mod.rs
@@ -1,0 +1,5 @@
+mod commands;
+mod print;
+pub mod setup;
+
+pub use commands::run;
